Main and readme were created!
In this example, we create a weighted graph and add vertices "A", "B", "C", "D", and "E". We then add edges with their corresponding weights to connect the vertices. Next, we create an instance of DijkstraSearch and BreadthFirstSearch classes, passing the graph as a parameter. Finally, we use the findShortestPath method to find the shortest path between vertex "A" and vertex "E" using both Dijkstra's algorithm and breadth-first search. The results are printed to the console.